a2a-reference-implementation/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangGraph & AI
langgraph==0.2.0
langchain==0.3.0
langchain-openai==0.2.0
openai==1.50.0

# OAuth2/JWT
python-jose[cryptography]==3.3.0
httpx==0.27.0
authlib==1.3.0

# Utilities
structlog==24.1.0
python-multipart==0.0.9
pyyaml==6.0.1

# A2A Protocol
a2a-sdk[http-server]>=0.1.0

# MCP Protocol
mcp>=1.0.0
